Fix Group comparisons and logic
Group used to inherit from list under the assumption buckets would be
stored in this underlying list, as such equality checks the underlying
list. However, all the code actually populates Group.buckets not the
underlying list directly.

So this removes the underlying list, and fixes the comparisons to
compare group.buckets.

Adds a hash function, and replaces the list of buckets with a
tuple of buckets, thus the group knows when these are modified.